renamed pipelinecomponents to processing
