+++ hrosailing/models/__init__.py
@@ -7,5 +7,8 @@ from hrosailing.models.influencemodel import (
     WindAngleCorrectingInfluenceModel
 )
 from hrosailing.models.data import Data
+from hrosailing.models.weather_model import (
+    WeatherModel, GriddedWeatherModel, NetCDFWeatherModel, OutsideGridException
+)
 
 import hrosailing.models.modelfunctions
