+++ hrosailing/models/weather_model.py
