+++ hrosailing/processing/constants.py
