+++ hrosailing/processing/interpolator.py
