+++ hrosailing/processing/regressor.py
