+++ hrosailing/processing/weigher.py
